+++ generate_sentic_dependency_graph.py
@@ -1,37 +1,15 @@
 # -*- coding: utf-8 -*-
-
 import numpy as np
 import spacy
-import pickle
-
+from common.fio import writefiles
 nlp = spacy.load('en_core_web_sm')
 
-
-def load_sentic_word():
-    """
-    load senticNet
-    """
-    path = './senticNet/senticnet_word.txt'
-    senticNet = {}
-    fp = open(path, 'r')
-    for line in fp:
-        line = line.strip()
-        if not line:
-            continue
-        word, sentic = line.split('\t')
-        senticNet[word] = sentic
-    fp.close()
-    return senticNet
-
-
+# customized function for graph_sentic_aspect_dep_adj_matrix
 def dependency_adj_matrix(text, aspect, senticNet):
     # https://spacy.io/docs/usage/processing-text
     document = nlp(text)
     seq_len = len(text.split())
     matrix = np.ones((seq_len, seq_len)).astype('float32')
-    #print('='*20+':')
-    #print(document)
-    #print(senticNet)
     
     for token in document:
         if str(token) in senticNet:
@@ -67,30 +45,5 @@ def dependency_adj_matrix(text, aspect, senticNet):
                       matrix[child.i][token.i] = 1 * sentic + s
     return matrix
 
-def process(filename):
-    senticNet = load_sentic_word()
-    fin = open(filename, 'r', encoding='utf-8', newline='\n', errors='ignore')
-    lines = fin.readlines()
-    fin.close()
-    idx2graph = {}
-    fout = open(filename+'.graph_sdat', 'wb')
-    for i in range(0, len(lines), 3):
-        text_left, _, text_right = [s.lower().strip() for s in lines[i].partition("$T$")]
-        aspect = lines[i + 1].lower().strip()
-        adj_matrix = dependency_adj_matrix(text_left+' '+aspect+' '+text_right, aspect, senticNet)
-        idx2graph[i] = adj_matrix
-    pickle.dump(idx2graph, fout)
-    print('done !!!'+filename)
-    fout.close() 
-
 if __name__ == '__main__':
-    process('./datasets/acl-14-short-data/train.raw')
-    process('./datasets/acl-14-short-data/test.raw')
-    process('./datasets/semeval14/restaurant_train.raw')
-    process('./datasets/semeval14/restaurant_test.raw')
-    process('./datasets/semeval14/laptop_train.raw')
-    process('./datasets/semeval14/laptop_test.raw')
-    process('./datasets/semeval15/restaurant_train.raw')
-    process('./datasets/semeval15/restaurant_test.raw')
-    process('./datasets/semeval16/restaurant_train.raw')
-    process('./datasets/semeval16/restaurant_test.raw')
+    writefiles(dependency_adj_matrix, suffix='.graph_sdat', graph_dep_type=3)
