+++ generate_sentic_graph.py
@@ -1,75 +1,6 @@
 # -*- coding: utf-8 -*-
+from common.fio import graph_sentic_aspect_dep_adj_matrix, writefiles
 
-import numpy as np
-#import spacy
-import pickle
-
-#nlp = spacy.load('en_core_web_sm')
-
-
-def load_sentic_word():
-    """
-    load senticNet
-    """
-    path = './senticNet/senticnet_word.txt'
-    senticNet = {}
-    fp = open(path, 'r')
-    for line in fp:
-        line = line.strip()
-        if not line:
-            continue
-        word, sentic = line.split('\t')
-        senticNet[word] = sentic
-    fp.close()
-    return senticNet
-
-
-def dependency_adj_matrix(text, aspect, senticNet):
-    word_list = text.split()
-    seq_len = len(word_list)
-    matrix = np.zeros((seq_len, seq_len)).astype('float32')
-    
-    for i in range(seq_len):
-        word = word_list[i]
-        if word in senticNet:
-            sentic = float(senticNet[word]) + 1.0
-        else:
-            sentic = 0
-        if word in aspect:
-            sentic += 1.0
-        for j in range(seq_len):
-            matrix[i][j] += sentic
-            matrix[j][i] += sentic
-    for i in range(seq_len):
-        if matrix[i][i] == 0:
-            matrix[i][i] = 1
-
-    return matrix
-
-def process(filename):
-    senticNet = load_sentic_word()
-    fin = open(filename, 'r', encoding='utf-8', newline='\n', errors='ignore')
-    lines = fin.readlines()
-    fin.close()
-    idx2graph = {}
-    fout = open(filename+'.sentic', 'wb')
-    for i in range(0, len(lines), 3):
-        text_left, _, text_right = [s.lower().strip() for s in lines[i].partition("$T$")]
-        aspect = lines[i + 1].lower().strip()
-        adj_matrix = dependency_adj_matrix(text_left+' '+aspect+' '+text_right, aspect, senticNet)
-        idx2graph[i] = adj_matrix
-    pickle.dump(idx2graph, fout)
-    print('done !!!', filename)
-    fout.close() 
-
+# dependency graph based on text, sentic net and sentiment aspects.
 if __name__ == '__main__':
-    process('./datasets/acl-14-short-data/train.raw')
-    process('./datasets/acl-14-short-data/test.raw')
-    process('./datasets/semeval14/restaurant_train.raw')
-    process('./datasets/semeval14/restaurant_test.raw')
-    process('./datasets/semeval14/laptop_train.raw')
-    process('./datasets/semeval14/laptop_test.raw')
-    process('./datasets/semeval15/restaurant_train.raw')
-    process('./datasets/semeval15/restaurant_test.raw')
-    process('./datasets/semeval16/restaurant_train.raw')
-    process('./datasets/semeval16/restaurant_test.raw')
+    writefiles(graph_sentic_aspect_dep_adj_matrix, suffix='.sentic', graph_dep_type=3)
